deletePage구현, write버그수정
페이지 삭제했을경우 대처, 중간글 삭제했을 경우 bnum다시 정렬, write로 게시글 작성후 새로고침하면 중복데이터 계속 삽입되는 버그 수정

+++ Java_JSP/JSPStudy_MVC_board/src/main/webapp/contentView.jsp
@@ -40,7 +40,7 @@
         <div class="post-actions">
           <a href="list.do" class="btn btn-outline">목록</a>
           <a href="modify.do" class="btn">수정</a>
-          <a href="delete.do" class="btn">삭제</a>
+          <a href="delete.do?bnum=${bDto.bnum}" class="btn" onclick="return confirm('정말로 이 게시글을 삭제하시겠습니까?')">삭제</a>
         </div>
       </footer>
     </article>
